requirements.txt
opencv-python
mediapipe
pycaw
pymongo
python-dotenv